Try locking issues after 300 days instead of 365
This will hopefully help reduce the the number of commends left on old closed tickets. (It becomes difficult to keep track of these comments since the ticket has been closed and many times the commends are not relevant in relation to the original issue. It is probably better for new issues to have their own tickets.)